% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data.R
\name{dl_data}
\alias{dl_data}
\title{Download drug effect size and Azimuth reference data.}
\usage{
dl_data(
  files = c("cmap_es_ind.qs", "l1000_drugs_es.qs", "l1000_genes_es.qs",
    "human_pbmc.qs"),
  check = FALSE
)
}
\arguments{
\item{files}{Character vector of file names to download.}

\item{check}{Check that existing data is loadable? Default
is FALSE.}
}
\value{
Downloads data into package folder.
}
\description{
Download drug effect size and Azimuth reference data.
}
\examples{

dl_data('example.qs')

}
