% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data.R
\name{load_data}
\alias{load_data}
\title{Load drug effect size and Azimuth reference data}
\usage{
load_data(
  file = c("cmap_es_ind.qs", "l1000_drugs_es.qs", "l1000_genes_es.qs", "human_pbmc.qs")
)
}
\arguments{
\item{file}{Character vector of drug effect size datasets to load. One of:
* `'cmap_es_ind.qs'` - CMAP02
* `'l1000_drugs_es.qs'` - L1000 compounds
* `'l1000_genes_es.qs'` - L1000 genetic perturbations
* `'human_pbmc.qs'` - Azimuth human PBMC reference}
}
\value{
data.frame of expression values. Rows are genes, columns are
  perturbations.
}
\description{
Downloads requested file if not done so previously.
}
\examples{

# dummy example (actual files are large)
load_drug_es("example.qs")
}
