% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_drug_es.R
\name{load_drug_es}
\alias{load_drug_es}
\title{Load drug effect size data}
\usage{
load_drug_es(
  file = c("cmap_es_ind.qs", "l1000_drugs_es.qs", "l1000_genes_es.qs")
)
}
\arguments{
\item{file}{Character vector of drug effect size datasets to load. One of
\code{'cmap_es_ind.qs'} (CMAP02),
\code{'l1000_drugs_es.qs'} (L1000 compounds), or
\code{'l1000_genes_es.qs'} (L1000 genetic perturbations).}
}
\value{
data.frame of expression values. Rows are genes, columns are
  perturbations.
}
\description{
Downloads requested file if not done so previously.
}
\examples{

# dummy example (actual files are large)
load_drug_es("example.qs")
}
