% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx2gene.R
\name{load_tx2gene}
\alias{load_tx2gene}
\title{Load transcript to gene map}
\usage{
load_tx2gene(species = "Homo sapiens", release = "94")
}
\arguments{
\item{species}{Character vector indicating species. Genus and species should
be space separated, not underscore. Default is \code{'Homo sapiens'}.}

\item{release}{EnsemblDB release. Should be same as used in
\code{\link[rkal]{build_kallisto_index}}.}
}
\value{
\code{data.frame} with columns \code{tx_id}, \code{gene_name}, and
  \code{entrezid}
}
\description{
Load transcript to gene map
}
\examples{

tx2gene <- load_tx2ene("Homo sapiens", "94")
}
